requirements.txt
google-cloud-speech==0.35.0